+++ clients/subaccount/CHANGELOG.md
@@ -1,5 +1,26 @@
 ### Changelog
 
+## 1.12.0 - 2026-08-24
+
+### Changed (4)
+
+- Added parameter `productType`
+  - affected methods:
+    - `getMovePositionHistoryForSubAccount()` (`GET /sapi/v1/sub-account/futures/move-position`)
+- Modified parameter `productType`:
+  - enum added: `OPTION`
+  - affected methods:
+    - `movePositionForSubAccount()` (`POST /sapi/v1/sub-account/futures/move-position`)
+- Modified response for `movePositionForSubAccount()` (`POST /sapi/v1/sub-account/futures/move-position`):
+  - `movePositionOrders`.items.`positionSide`: nullable `false` → `true`
+  - `movePositionOrders`.items.`positionSide`: nullable `false` → `true`
+
+- Modified response field `movePositionOrders`:
+  - items.`positionSide`: nullable `false` → `true`
+  - items.`positionSide`: nullable `false` → `true`
+  - affected events:
+    - `movePositionForSubAccountResponse`
+
 ## 1.11.0 - 2026-08-19
 
 ### Changed (2)